Changed buildId type to Int64 (#5054)
Changed buildId type to Int64 to correspond with Build.id type

+++ dashboard/lib/service/cocoon.dart
@@ -5,6 +5,7 @@
 import 'package:cocoon_common/rpc_model.dart';
 import 'package:cocoon_common/task_status.dart';
 import 'package:cocoon_integration_test/cocoon_integration_test.dart';
+import 'package:fixnum/fixnum.dart';
 import 'package:flutter/foundation.dart';
 
 import 'appengine_cocoon.dart';
@@ -87,7 +88,7 @@ abstract class CocoonService {
     required String? idToken,
     required String repo,
     required int pr,
-    required int buildId,
+    required Int64 buildId,
     String owner = 'flutter',
   });
